Add files via upload
Preliminary studies about the qualities of advisory comments. This includes studying correlations of the feature values such as sentiment analysis, polarity analysis and cosine similarity between the post asking for advice and the advisory comment given as a response (cosine similarity measured with word2vec and doc2vec), and the comment's score. The score is a measure of the quality of that advice, in other words, how good this advice is.

First download the data, then calculate the features, then calculate the correlations.